feat(Rover .ha files): Now able to extract NAVCAM 1024x1024 images from logs.
All LDT transfers are now extracted and saved to disk. If not a PanCam LDT they are placed within the LDT_RAW folder. If the LDT_RAW file is recognised as a NAVCAM image by the size of the file, a pgm file is created within the NAVCAM folder.

Also created NAVCAM.md to document structure and pgm header.