(swis.hs_extract): Now ignores lines that are not the correct size, set to allow 15 consecutive errors before an error is generated.
When running the SWIS simulations through Docker lines in the text log can be concatanted or appended to a previous line. This causes errors when parsing the HS data, now HS will ignore lines of the wrong length and raise an error if more than 15 lines in a row are corrupted.